pyproject.toml
[project]
name = "github-issues-linear"
version = "0.1.0"
description = "Sync GitHub issues to Linear"
authors = [{ name = "Ola Austine", email = "austineolanrewaju@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "pre-commit>=4.3.0",
    "pydantic>=2.12.0",
    "pygithub>=2.8.1",
    "pytest>=8.4.2",
    "redis>=6.4.0",
]
